# api/openapi.yaml
openapi: 3.0.3
info:
  title: Messaggio
  description: Микросервис обработки сообщений
  version: 0.0.1
  contact:
    email: lazylex@mail.ru
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: 'http://localhost:8897'
tags:
  - name: messages
    description: Работа с сообщениями
paths:
  /msg:
    post:
      tags:
        - messages
      summary: Отправка сообщения в сервис
      description: Отправка сообщения в сервис, сохранение в БД и отправка в брокер сообщений
      operationId: ProcessMessage
      requestBody:
        required: true
        content:
          text/plain:
            example: just simple message
      responses:
        '102':
          description: Сообщение принято в сервисе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusID'
        '400':
          description: Ошибка чтения или пустое тело сообщения
        '500':
          description: Внутренняя ошибка сервера
  /statistic:
    get:
      tags:
        - messages
      summary: Получение статистики по отправленным и ожидающим сообщениям
      description: Возвращает JSON со статистикой по сообщениям с момента запуска приложения. Содержит данные об общем 
        количестве пришедших на обработку сообщений (total), несвоевременно обновленных в БД статусах 
        (statuses_sent_to_outbox), несвоевременно сохраненных в БД сообщениях (messages_sent_to_outbox)
      operationId: Statistic
      responses:
        '200':
          description: Успешное получение статистики
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Statistic'
        '500':
          description: Внутренняя ошибка сервера
  /processed-statistic:
    get:
      tags:
        - messages
      summary: Получение статистики по обработанным сообщениям
      description: Возвращает JSON со статистикой по обработанным за последний час, день, неделю, месяц сообщениям
      operationId: ProcessedStatistic
      responses:
        '200':
          description: Успешное получение статистики
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessedStatistic'
        '500':
          description: Внутренняя ошибка сервера
          
components:
  schemas:
    StatusID:
      type: object
      description: Статус и идентификатор сообщения
      properties:
        status:
          type: string
          description: Статус сообщения
          example: saved, sent to the broker...
        id:
          type: string
          format: uuid
          description: Идентификатор отправленного сообщения
          example: cb0e57e2-5050-4644-8ada-1dc23ef1f518
    Statistic:
      type: object
      description: Статистика обработки сообщений
      properties:
        total:
          type: integer
          description: Всего получено сообщений на обработку
          example: 1569
          minimum: 0
        messages_sent_to_outbox:
          type: integer
          description: Всего было сохранено сообщений во временное хранилище для дальнейших попыток сохранения в БД
          example: 3
          minimum: 0
        messages_returned_from_outbox:
          type: integer
          description: Всего удалось переместить сообщений из outbox в БД
          example: 3
          minimum: 0
    ProcessedStatistic:
      type: object
      description: Статистика обработки сообщений
      properties:
        in_hour:
          type: integer
          description: Обработано сообщений за час
          example: 15
          minimum: 0
        in_day:
          type: integer
          description: Обработано сообщений за день
          example: 30
          minimum: 0
        in_week:
          type: integer
          description: Обработано сообщений за неделю
          example: 9000
          minimum: 0
        in_month:
          type: integer
          description: Обработано сообщений за месяц
          example: 2700000
          minimum: 0